requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.4.1
blinker==1.8.2
boto3==1.34.117
botocore==1.34.117
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deepl==1.18.0
docx==0.2.4
easyocr==1.7.1
elevenlabs==1.2.2
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
gTTS==2.5.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.1
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
lazy_loader==0.4
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauth2client==4.1.3
opencv-python-headless==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.2.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.1
reportlab==4.2.0
requests==2.32.2
rsa==4.9
s3transfer==0.10.1
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tifffile==2024.5.10
torch==2.3.0
torchvision==0.18.0
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
